Add NixOS Packages
Added a NixOS Package for Illogical Impulse, and have it installed
automatically upon enabling the home-manager module's enable option.
There is an issue with pkgconf not recognizing gtksourceviewmm as being
of version 3.0, and it isn't trivial to use an older package version
because the last version which matches the package used in Arch is from
before nixpkgs had a flake.nix - so it cannot just be added as a flake
input. For this reason MicroTeX has not been packaged at this time.
MicroTeX will require more testing to package - perhaps using a custom
gtksourceviewmm package with the older version?